Modularized layout files
